backend/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
python-dotenv
openpyxl
aiogram==3.4.1